mvd product db ipcmain handle to seperate files
